+++ Source/Platform/Android/Content/Shader/ShadowMap/PointLightShadow.vert
@@ -1,5 +1,6 @@
 #version 320 es
-precision mediump float;
+
+precision highp float;
 layout (location = 0) in vec3 position;
 
 uniform mat4 ModelTransform;
